Update estimated year of birth
To the questions if 0,1 (fawns) 1,2 (yearlings) and only 2 (adults) years are possible as difference between capture and year of birth I adjusted the queries. For example for fawns, I think both 0 and 1 year difference are possible but it depends from the moment of capture. If the first capture was before April than the year of birth should be 1 year before (1) since fawns are born after April (May/June). If the first capture was after April than the year of birth should be the same as the year of capture (0).